+++ level.go
@@ -1,6 +1,8 @@
 package golog
 
 import (
+	"fmt"
+	"strconv"
 	"strings"
 
 	"github.com/kataras/pio"
@@ -8,6 +10,25 @@ import (
 
 // Level is a number which defines the log level.
 type Level uint32
+
+// MarshalJSON implements the json marshaler for Level.
+func (l Level) MarshalJSON() ([]byte, error) {
+	if level, ok := Levels[l]; ok {
+		return []byte(strconv.Quote(level.Name)), nil
+	}
+
+	return nil, fmt.Errorf("unknown level %v", l)
+}
+
+// String implements the fmt.Stringer interface for level.
+// Returns the level's name.
+func (l Level) String() string {
+	if level, ok := Levels[l]; ok {
+		return level.Name
+	}
+
+	return ""
+}
 
 // The available built'n log levels, users can add or modify a level via `Levels` field.
 const (
@@ -35,50 +56,55 @@ var Levels = map[Level]*LevelMetadata{
 	DisableLevel: {
 		Name:             "disable",
 		AlternativeNames: []string{"disabled"},
-		RawText:          "",
-		ColorfulText:     "",
+		Title:            "",
 	},
 	FatalLevel: {
-		Name:    "fatal",
-		RawText: "[FTAL]",
-		// white foreground but red background, it's nice
-		ColorfulText: pio.RedBackground("[FTAL]"),
+		Name:      "fatal",
+		Title:     "[FTAL]",
+		ColorCode: pio.Red,
+		Style:     []pio.RichOption{pio.Background},
 	},
 	ErrorLevel: {
-		Name:         "error",
-		RawText:      "[ERRO]",
-		ColorfulText: pio.Red("[ERRO]"),
+		Name:      "error",
+		Title:     "[ERRO]",
+		ColorCode: pio.Red,
 	},
 	WarnLevel: {
 		Name:             "warn",
 		AlternativeNames: []string{"warning"},
-		RawText:          "[WARN]",
-		ColorfulText:     pio.Purple("[WARN]"),
+		Title:            "[WARN]",
+		ColorCode:        pio.Magenta,
 	},
 	InfoLevel: {
-		Name:         "info",
-		RawText:      "[INFO]",
-		ColorfulText: pio.LightGreen("[INFO]"),
+		Name:      "info",
+		Title:     "[INFO]",
+		ColorCode: pio.Cyan,
 	},
 	DebugLevel: {
-		Name:         "debug",
-		RawText:      "[DBUG]",
-		ColorfulText: pio.Yellow("[DBUG]"),
+		Name:      "debug",
+		Title:     "[DBUG]",
+		ColorCode: pio.Yellow,
 	},
 }
 
-func fromLevelName(levelName string) Level {
+// ParseLevel returns a `golog.Level` from a string level.
+// Note that all existing log levels (name, prefix and color) can be customized
+// and new one can be added by the package-level `golog.Levels` map variable.
+func ParseLevel(levelName string) Level {
+	levelName = strings.ToLower(levelName)
+
 	for level, meta := range Levels {
-		if meta.Name == levelName {
+		if strings.ToLower(meta.Name) == levelName {
 			return level
 		}
 
 		for _, altName := range meta.AlternativeNames {
-			if altName == levelName {
+			if strings.ToLower(altName) == levelName {
 				return level
 			}
 		}
 	}
+
 	return DisableLevel
 }
 
@@ -95,19 +121,14 @@ type LevelMetadata struct {
 	// AlternativeNames = []string{"warning"}, it's an optional field,
 	// therefore we keep Name as a simple string and created this new field.
 	AlternativeNames []string
-	// Tha RawText will be the prefix of the log level
-	// when output doesn't supports colors.
-	//
-	// When RawText is changed its ColorfulText is also changed
-	// to a default color, but callers are able to change it too.
-	RawText string
-	// The ColorfulText will be the prefix of the log level
-	// when output supports colors, almost everything except
-	// os files and putty-based terminals(?).
-	//
-	// If ColorfulText is empty then built'n colors
-	// are being used to wrap the "RawText".
-	ColorfulText string
+	// Tha Title is the prefix of the log level.
+	// See `ColorCode` and `Style` too.
+	// Both `ColorCode` and `Style` should be respected across writers.
+	Title string
+	// ColorCode a color for the `Title`.
+	ColorCode int
+	// Style one or more rich options for the `Title`.
+	Style []pio.RichOption
 }
 
 // Text returns the text that should be
@@ -115,27 +136,17 @@ type LevelMetadata struct {
 // log level is being written.
 func (m *LevelMetadata) Text(enableColor bool) string {
 	if enableColor {
-		return m.ColorfulText
+		return pio.Rich(m.Title, m.ColorCode, m.Style...)
 	}
-	return m.RawText
+	return m.Title
 }
 
 // SetText can modify the prefix that will be prepended
 // to the output message log when `Error/Errorf` functions are being used.
-//
-// If "newRawText" is empty then it will just skip the Text set-ing.
-// If "newColorfulText" is empty then it will update the text color version using
-// the default values by using the new raw text.
-func (m *LevelMetadata) SetText(newRawText string, newColorfulText string) {
-	if newRawText != "" {
-		oldRawText := m.RawText
-		m.RawText = newRawText
-		m.ColorfulText = strings.Replace(m.ColorfulText, oldRawText, newRawText, -1)
-	}
-	if newColorfulText != "" {
-		m.ColorfulText = newColorfulText
-	}
-
+func (m *LevelMetadata) SetText(title string, colorCode int, style ...pio.RichOption) {
+	m.Title = title
+	m.ColorCode = colorCode
+	m.Style = style
 }
 
 var (
@@ -193,6 +204,16 @@ var (
 	GetTextForLevel = func(level Level, enableColor bool) string {
 		if meta, ok := Levels[level]; ok {
 			return meta.Text(enableColor)
+		}
+		return ""
+	}
+	
+	// GetNameForLevel is the function which
+	// has the "final" responsibility to generagte the name of the level
+	// that is prepended to the leveled log message
+	GetNameForLevel = func(level Level) string {
+		if meta, ok := Levels[level]; ok {
+			return meta.Name
 		}
 		return ""
 	}