+++ logger.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"io"
 	"os"
+	"strings"
 	"sync"
 	"time"
 
@@ -26,9 +27,11 @@ type Handler func(value *Log) (handled bool)
 
 // Logger is our golog.
 type Logger struct {
-	Prefix     []byte
+	Prefix     string
 	Level      Level
 	TimeFormat string
+	// Limit stacktrace entries on `Debug` level.
+	StacktraceLimit int
 	// if new line should be added on all log functions, even the `F`s.
 	// It defaults to true.
 	//
@@ -37,9 +40,16 @@ type Logger struct {
 	// Note that this will not override the time and level prefix,
 	// if you want to customize the log message please read the examples
 	// or navigate to: https://github.com/kataras/golog/issues/3#issuecomment-355895870.
-	NewLine  bool
-	mu       sync.Mutex
-	Printer  *pio.Printer
+	NewLine bool
+	mu      sync.RWMutex // for logger field changes and printing through pio hijacker.
+	Printer *pio.Printer
+	// The per log level raw writers, optionally.
+	LevelOutput map[Level]io.Writer
+
+	formatters     map[string]Formatter // available formatters.
+	formatter      Formatter            // the current formatter for all logs.
+	LevelFormatter map[Level]Formatter  // per level formatter.
+
 	handlers []Handler
 	once     sync.Once
 	logs     sync.Pool
@@ -50,26 +60,43 @@ type Logger struct {
 // and level to `InfoLevel`.
 func New() *Logger {
 	return &Logger{
-		Level:      InfoLevel,
-		TimeFormat: "2006/01/02 15:04",
-		NewLine:    true,
-		Printer:    pio.NewPrinter("", os.Stdout).EnableDirectOutput().Hijack(logHijacker),
-		children:   newLoggerMap(),
+		Level:       InfoLevel,
+		TimeFormat:  "2006/01/02 15:04",
+		NewLine:     true,
+		Printer:     pio.NewPrinter("", os.Stdout).EnableDirectOutput().Hijack(logHijacker).SetSync(true),
+		LevelOutput: make(map[Level]io.Writer),
+		formatters: map[string]Formatter{ // the available builtin formatters.
+			"json": new(JSONFormatter),
+		},
+		LevelFormatter: make(map[Level]Formatter),
+		children:       newLoggerMap(),
 	}
 }
 
+// Fields is a map type.
+// One or more values of `Fields` type can be passed
+// on all Log methods except `Print/Printf/Println` to set the `Log.Fields` field,
+// which can be accessed through a custom LogHandler.
+type Fields map[string]interface{}
+
 // acquireLog returns a new log fom the pool.
-func (l *Logger) acquireLog(level Level, msg string, withPrintln bool) *Log {
+func (l *Logger) acquireLog(level Level, msg string, withPrintln bool, fields Fields) *Log {
 	log, ok := l.logs.Get().(*Log)
 	if !ok {
 		log = &Log{
 			Logger: l,
 		}
 	}
+
 	log.NewLine = withPrintln
-	log.Time = time.Now()
+	if l.TimeFormat != "" {
+		log.Time = time.Now()
+		log.Timestamp = log.Time.Unix()
+	}
 	log.Level = level
 	log.Message = msg
+	log.Fields = fields
+	log.Stacktrace = log.Stacktrace[:0]
 	return log
 }
 
@@ -77,6 +104,8 @@ func (l *Logger) acquireLog(level Level, msg string, withPrintln bool) *Log {
 func (l *Logger) releaseLog(log *Log) {
 	l.logs.Put(log)
 }
+
+var spaceBytes = []byte(" ")
 
 // we could use marshal inside Log but we don't have access to printer,
 // we could also use the .Handle with NopOutput too but
@@ -88,25 +117,41 @@ var logHijacker = func(ctx *pio.Ctx) {
 		return
 	}
 
-	line := GetTextForLevel(l.Level, ctx.Printer.IsTerminal)
-	if line != "" {
-		line += " "
+	logger := l.Logger
+	logger.mu.Lock()
+	defer logger.mu.Unlock()
+
+	w := logger.getOutput(l.Level)
+	if f := logger.getFormatter(); f != nil {
+		if f.Format(w, l) {
+			ctx.Store(nil, pio.ErrHandled)
+			return
+		}
+	}
+
+	if l.Level != DisableLevel {
+		if level, ok := Levels[l.Level]; ok {
+			pio.WriteRich(w, level.Title, level.ColorCode, level.Style...)
+			w.Write(spaceBytes)
+		}
 	}
 
 	if t := l.FormatTime(); t != "" {
-		line += t + " "
-	}
-	line += l.Message
-
-	var b []byte
-	if pref := l.Logger.Prefix; len(pref) > 0 {
-		b = append(pref, []byte(line)...)
-	} else {
-		b = []byte(line)
+		fmt.Fprint(w, t)
+		w.Write(spaceBytes)
 	}
 
-	ctx.Store(b, nil)
-	ctx.Next()
+	if prefix := logger.Prefix; len(prefix) > 0 {
+		fmt.Fprintf(w, prefix)
+	}
+
+	fmt.Fprint(w, l.Message)
+
+	if logger.NewLine {
+		fmt.Fprintln(w)
+	}
+
+	ctx.Store(nil, pio.ErrHandled)
 }
 
 // NopOutput disables the output.
@@ -133,14 +178,13 @@ func (l *Logger) AddOutput(writers ...io.Writer) *Logger {
 
 // SetPrefix sets a prefix for this "l" Logger.
 //
-// The prefix is the first space-separated
-// word that is being presented to the output.
-// It's written even before the log level text.
+// The prefix is the text that is being presented
+// to the output right before the log's message.
 //
 // Returns itself.
 func (l *Logger) SetPrefix(s string) *Logger {
 	l.mu.Lock()
-	l.Prefix = []byte(s)
+	l.Prefix = s
 	l.mu.Unlock()
 	return l
 }
@@ -157,6 +201,18 @@ func (l *Logger) SetTimeFormat(s string) *Logger {
 	return l
 }
 
+// SetStacktraceLimit sets a stacktrace entries limit
+// on `Debug` level.
+// Zero means all number of stack entries will be logged.
+// Negative value disables the stacktrace field.
+func (l *Logger) SetStacktraceLimit(limit int) *Logger {
+	l.mu.Lock()
+	l.StacktraceLimit = limit
+	l.mu.Unlock()
+
+	return l
+}
+
 // DisableNewLine disables the new line suffix on every log function, even the `F`'s,
 // the caller should add "\n" to the log message manually after this call.
 //
@@ -167,6 +223,84 @@ func (l *Logger) DisableNewLine() *Logger {
 	l.mu.Unlock()
 
 	return l
+}
+
+// RegisterFormatter registers a Formatter for this logger.
+func (l *Logger) RegisterFormatter(f Formatter) *Logger {
+	l.mu.Lock()
+	l.formatters[f.String()] = f
+	l.mu.Unlock()
+	return l
+}
+
+// SetFormat sets a formatter for all logger's logs.
+func (l *Logger) SetFormat(formatter string, opts ...interface{}) *Logger {
+	l.mu.RLock()
+	f, ok := l.formatters[formatter]
+	l.mu.RUnlock()
+
+	if ok {
+		l.mu.Lock()
+		l.formatter = f.Options(opts...)
+		l.mu.Unlock()
+	}
+
+	return l
+}
+
+// SetLevelFormat changes the output format for the given "levelName".
+func (l *Logger) SetLevelFormat(levelName string, formatter string, opts ...interface{}) *Logger {
+	l.mu.RLock()
+	f, ok := l.formatters[formatter]
+	l.mu.RUnlock()
+
+	if ok {
+		l.mu.Lock()
+		l.LevelFormatter[ParseLevel(levelName)] = f.Options(opts...)
+		l.mu.Unlock()
+	}
+
+	return l
+}
+
+func (l *Logger) getFormatter() Formatter {
+	f, ok := l.LevelFormatter[l.Level]
+	if !ok {
+		f = l.formatter
+	}
+
+	if f == nil {
+		return nil
+	}
+
+	return f
+}
+
+// SetLevelOutput sets a destination log output for the specific "levelName".
+// For multiple writers use the `io.Multiwriter` wrapper.
+func (l *Logger) SetLevelOutput(levelName string, w io.Writer) *Logger {
+	l.mu.Lock()
+	l.LevelOutput[ParseLevel(levelName)] = w
+	l.mu.Unlock()
+	return l
+}
+
+// GetLevelOutput returns the responsible writer for the given "levelName".
+// If not a registered writer is set for that level then it returns
+// the logger's default printer. It does NOT return nil.
+func (l *Logger) GetLevelOutput(levelName string) io.Writer {
+	l.mu.RLock()
+	w := l.getOutput(ParseLevel(levelName))
+	l.mu.RUnlock()
+	return w
+}
+
+func (l *Logger) getOutput(level Level) io.Writer {
+	w, ok := l.LevelOutput[level]
+	if !ok {
+		w = l.Printer
+	}
+	return w
 }
 
 // SetLevel accepts a string representation of
@@ -185,18 +319,21 @@ func (l *Logger) DisableNewLine() *Logger {
 // Returns itself.
 func (l *Logger) SetLevel(levelName string) *Logger {
 	l.mu.Lock()
-	l.Level = fromLevelName(levelName)
+	l.Level = ParseLevel(levelName)
 	l.mu.Unlock()
 
 	return l
 }
 
-func (l *Logger) print(level Level, msg string, newLine bool) {
+func (l *Logger) print(level Level, msg string, newLine bool, fields Fields) {
 	if l.Level >= level {
 		// newLine passed here in order for handler to know
 		// if this message derives from Println and Leveled functions
 		// or by simply, Print.
-		log := l.acquireLog(level, msg, newLine)
+		log := l.acquireLog(level, msg, newLine, fields)
+		if level == DebugLevel {
+			log.Stacktrace = GetStacktrace(l.StacktraceLimit)
+		}
 		// if not handled by one of the handler
 		// then print it as usual.
 		if !l.handled(log) {
@@ -217,33 +354,67 @@ func (l *Logger) print(level Level, msg string, newLine bool) {
 
 // Print prints a log message without levels and colors.
 func (l *Logger) Print(v ...interface{}) {
-	l.print(DisableLevel, fmt.Sprint(v...), l.NewLine)
+	l.print(DisableLevel, fmt.Sprint(v...), l.NewLine, nil)
 }
 
 // Printf formats according to a format specifier and writes to `Printer#Output` without levels and colors.
 func (l *Logger) Printf(format string, args ...interface{}) {
-	l.print(DisableLevel, fmt.Sprintf(format, args...), l.NewLine)
+	l.print(DisableLevel, fmt.Sprintf(format, args...), l.NewLine, nil)
 }
 
 // Println prints a log message without levels and colors.
 // It adds a new line at the end, it overrides the `NewLine` option.
 func (l *Logger) Println(v ...interface{}) {
-	l.print(DisableLevel, fmt.Sprint(v...), true)
+	l.print(DisableLevel, fmt.Sprint(v...), true, nil)
+}
+
+func splitArgsFields(values []interface{}) ([]interface{}, Fields) {
+	var (
+		args   = values[:0]
+		fields Fields
+	)
+
+	for _, value := range values {
+		if f, ok := value.(Fields); ok {
+			if fields == nil {
+				fields = make(Fields)
+			}
+
+			for k, v := range f {
+				fields[k] = v
+			}
+
+			continue
+		}
+
+		args = append(args, value) // use it as fmt argument.
+	}
+
+	return args, fields
 }
 
 // Log prints a leveled log message to the output.
 // This method can be used to use custom log levels if needed.
 // It adds a new line in the end.
 func (l *Logger) Log(level Level, v ...interface{}) {
-	l.print(level, fmt.Sprint(v...), l.NewLine)
+	if l.Level >= level {
+		args, fields := splitArgsFields(v)
+		l.print(level, fmt.Sprint(args...), l.NewLine, fields)
+	}
 }
 
 // Logf prints a leveled log message to the output.
 // This method can be used to use custom log levels if needed.
 // It adds a new line in the end.
 func (l *Logger) Logf(level Level, format string, args ...interface{}) {
-	msg := fmt.Sprintf(format, args...)
-	l.Log(level, msg)
+	if l.Level >= level {
+		arguments, fields := splitArgsFields(args)
+		msg := format
+		if len(arguments) > 0 {
+			msg = fmt.Sprintf(msg, arguments...)
+		}
+		l.print(level, msg, l.NewLine, fields)
+	}
 }
 
 // Fatal `os.Exit(1)` exit no matter the level of the logger.
@@ -257,8 +428,7 @@ func (l *Logger) Fatal(v ...interface{}) {
 // If the logger's level is fatal, error, warn, info or debug
 // then it will print the log message too.
 func (l *Logger) Fatalf(format string, args ...interface{}) {
-	msg := fmt.Sprintf(format, args...)
-	l.Fatal(msg)
+	l.Logf(FatalLevel, format, args...)
 }
 
 // Error will print only when logger's Level is error, warn, info or debug.
@@ -268,8 +438,7 @@ func (l *Logger) Error(v ...interface{}) {
 
 // Errorf will print only when logger's Level is error, warn, info or debug.
 func (l *Logger) Errorf(format string, args ...interface{}) {
-	msg := fmt.Sprintf(format, args...)
-	l.Error(msg)
+	l.Logf(ErrorLevel, format, args...)
 }
 
 // Warn will print when logger's Level is warn, info or debug.
@@ -279,8 +448,14 @@ func (l *Logger) Warn(v ...interface{}) {
 
 // Warnf will print when logger's Level is warn, info or debug.
 func (l *Logger) Warnf(format string, args ...interface{}) {
-	msg := fmt.Sprintf(format, args...)
-	l.Warn(msg)
+	l.Logf(WarnLevel, format, args...)
+}
+
+// Warningf exactly the same as `Warnf`.
+// It's here for badger integration:
+// https://github.com/dgraph-io/badger/blob/ef28ef36b5923f12ffe3a1702bdfa6b479db6637/logger.go#L25
+func (l *Logger) Warningf(format string, args ...interface{}) {
+	l.Warnf(format, args...)
 }
 
 // Info will print when logger's Level is info or debug.
@@ -290,8 +465,7 @@ func (l *Logger) Info(v ...interface{}) {
 
 // Infof will print when logger's Level is info or debug.
 func (l *Logger) Infof(format string, args ...interface{}) {
-	msg := fmt.Sprintf(format, args...)
-	l.Info(msg)
+	l.Logf(InfoLevel, format, args...)
 }
 
 // Debug will print when logger's Level is debug.
@@ -301,13 +475,7 @@ func (l *Logger) Debug(v ...interface{}) {
 
 // Debugf will print when logger's Level is debug.
 func (l *Logger) Debugf(format string, args ...interface{}) {
-	// On debug mode don't even try to fmt.Sprintf if it's not required,
-	// this can be used to allow `Debugf` to be called without even the `fmt.Sprintf`'s
-	// performance cost if the logger doesn't allow debug logging.
-	if l.Level >= DebugLevel {
-		msg := fmt.Sprintf(format, args...)
-		l.Debug(msg)
-	}
+	l.Logf(DebugLevel, format, args...)
 }
 
 // Install receives  an external logger
@@ -399,8 +567,8 @@ func (l *Logger) Scan(r io.Reader) (cancel func()) {
 				return nil, pio.ErrMarshalNotResponsible
 			}
 
-			formattedTime := time.Now().Format(l.TimeFormat)
-			if formattedTime != "" {
+			if l.TimeFormat != "" {
+				formattedTime := time.Now().Format(l.TimeFormat)
 				line = append([]byte(formattedTime+" "), line...)
 			}
 
@@ -414,57 +582,135 @@ func (l *Logger) Scan(r io.Reader) (cancel func()) {
 // Clone returns a copy of this "l" Logger.
 // This copy is returned as pointer as well.
 func (l *Logger) Clone() *Logger {
+	// copy level output and format maps.
+	formats := make(map[string]Formatter, len(l.formatters))
+	for k, v := range l.formatters {
+		formats[k] = v
+	}
+	levelFormat := make(map[Level]Formatter, len(l.LevelFormatter))
+	for k, v := range l.LevelFormatter {
+		levelFormat[k] = v
+	}
+	levelOutput := make(map[Level]io.Writer, len(l.LevelOutput))
+	for k, v := range l.LevelOutput {
+		levelOutput[k] = v
+	}
+
 	return &Logger{
-		Prefix:     l.Prefix,
-		Level:      l.Level,
-		TimeFormat: l.TimeFormat,
-		NewLine:    l.NewLine,
-		Printer:    l.Printer,
-		handlers:   l.handlers,
-		children:   newLoggerMap(),
-		mu:         sync.Mutex{},
-		once:       sync.Once{},
+		Prefix:         l.Prefix,
+		Level:          l.Level,
+		TimeFormat:     l.TimeFormat,
+		NewLine:        l.NewLine,
+		Printer:        l.Printer,
+		LevelOutput:    levelOutput,
+		formatter:      l.formatter,
+		formatters:     formats,
+		LevelFormatter: levelFormat,
+		handlers:       l.handlers,
+		children:       newLoggerMap(),
+		mu:             sync.RWMutex{},
+		once:           sync.Once{},
 	}
 }
 
 // Child (creates if not exists and) returns a new child
-// Logger based on the "l"'s fields.
+// Logger based on the current logger's fields.
 //
 // Can be used to separate logs by category.
-func (l *Logger) Child(name string) *Logger {
-	return l.children.getOrAdd(name, l)
+// If the "key" is string then it's used as prefix,
+// which is appended to the current prefix one.
+func (l *Logger) Child(key interface{}) *Logger {
+	return l.children.getOrAdd(key, l)
+}
+
+// SetChildPrefix same as `SetPrefix` but it does NOT
+// override the existing, instead the given "prefix"
+// is appended to the current one. It's useful
+// to chian loggers with their own names/prefixes.
+// It does add the ": " in the end of "prefix" if it's missing.
+// It returns itself.
+func (l *Logger) SetChildPrefix(prefix string) *Logger {
+	if prefix == "" {
+		return l
+	}
+
+	// if prefix doesn't end with a whitespace, then add it here.
+	if !strings.HasSuffix(prefix, ": ") {
+		prefix += ": "
+	}
+
+	l.mu.Lock()
+	if l.Prefix != "" {
+		if !strings.HasSuffix(l.Prefix, " ") {
+			l.Prefix += " "
+		}
+	}
+	l.Prefix += prefix
+	l.mu.Unlock()
+
+	return l
+}
+
+// LastChild returns the last registered child Logger.
+func (l *Logger) LastChild() *Logger {
+	return l.children.getLast()
 }
 
 type loggerMap struct {
-	mu    sync.RWMutex
-	Items map[string]*Logger
+	mu           sync.RWMutex
+	Items        map[interface{}]*Logger
+	itemsOrdered map[int]interface{} // registration order of logger and its key.
 }
 
 func newLoggerMap() *loggerMap {
 	return &loggerMap{
-		Items: make(map[string]*Logger),
+		Items:        make(map[interface{}]*Logger),
+		itemsOrdered: make(map[int]interface{}),
 	}
 }
 
-func (m *loggerMap) getOrAdd(name string, parent *Logger) *Logger {
+func (m *loggerMap) getByIndex(index int) (l *Logger) {
 	m.mu.RLock()
-	logger, ok := m.Items[name]
+	if key, ok := m.itemsOrdered[index]; ok {
+		l = m.Items[key]
+	}
+	m.mu.RUnlock()
+
+	return l
+}
+
+func (m *loggerMap) getLast() *Logger {
+	m.mu.RLock()
+	n := len(m.Items)
+	m.mu.RUnlock()
+	if n == 0 {
+		return nil
+	}
+
+	return m.getByIndex(n - 1)
+}
+
+func (m *loggerMap) getOrAdd(key interface{}, parent *Logger) *Logger {
+	m.mu.RLock()
+	logger, ok := m.Items[key]
 	m.mu.RUnlock()
 	if ok {
 		return logger
 	}
 
 	logger = parent.Clone()
-	prefix := name
-
-	// if prefix doesn't end with a whitespace, then add it here.
-	if lb := name[len(prefix)-1]; lb != ' ' {
-		prefix += ": "
+	childPrefix := ""
+	switch v := key.(type) {
+	case string:
+		childPrefix = v
+	case fmt.Stringer:
+		childPrefix = v.String()
 	}
+	logger.SetChildPrefix(childPrefix)
 
-	logger.SetPrefix(prefix)
 	m.mu.Lock()
-	m.Items[name] = logger
+	m.itemsOrdered[len(m.Items)] = key
+	m.Items[key] = logger
 	m.mu.Unlock()
 
 	return logger